RpcConfigFactory: move to eCampApi\V1\ namespace
Issue: #631
