+++ backend/module/eCampApi/config/Rpc/index.config.php
@@ -1,7 +1,7 @@
 <?php
 
-use eCampApi\RpcConfigFactory;
 use eCampApi\V1\Rpc\Index\IndexController;
+use eCampApi\V1\RpcConfigFactory;
 
 return RpcConfigFactory::forRoute('e-camp-api.rpc.index')
     ->setController(IndexController::class)
