+++ backend/module/eCampApi/config/Rpc/printer.config.php
@@ -1,7 +1,7 @@
 <?php
 
-use eCampApi\RpcConfigFactory;
 use eCampApi\V1\Rpc\Printer\PrinterController;
+use eCampApi\V1\RpcConfigFactory;
 
 return RpcConfigFactory::forRoute('e-camp-api.rpc.printer')
     ->setController(PrinterController::class)
