+++ backend/module/eCampApi/config/Rpc/profile.config.php
@@ -1,7 +1,7 @@
 <?php
 
-use eCampApi\RpcConfigFactory;
 use eCampApi\V1\Rpc\Profile\ProfileController;
+use eCampApi\V1\RpcConfigFactory;
 
 return RpcConfigFactory::forRoute('e-camp-api.rpc.profile')
     ->setController(ProfileController::class)
