+++ backend/module/eCampApi/config/Rpc/register.config.php
@@ -1,7 +1,7 @@
 <?php
 
-use eCampApi\RpcConfigFactory;
 use eCampApi\V1\Rpc\Register\RegisterController;
+use eCampApi\V1\RpcConfigFactory;
 
 return RpcConfigFactory::forRoute('e-camp-api.rpc.register')
     ->setController(RegisterController::class)
