+++ backend/module/eCampApi/src/eCampApi/V1/RpcConfigFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace eCampApi;
+namespace eCampApi\V1;
 
 use eCamp\Lib\Entity\EntityLink;
 use eCamp\Lib\Entity\EntityLinkCollection;
